refactor: Renamed Connectors to SemanticKernel.
