+++ LangChain.sln
@@ -371,9 +371,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Milvus", "Milvus", "{3CB20A
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LangChain.Databases.Milvus", "src\Databases\Milvus\src\LangChain.Databases.Milvus.csproj", "{259C2E5D-6AA9-4384-8F55-24C4D4F1CF18}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Connectors", "Connectors", "{FB278634-1B3C-4AF1-BFC1-BCC7E07D1388}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SemanticKernel", "SemanticKernel", "{3F426832-2CF3-4921-819F-13342F69A6E6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LangChain.Databases.Connectors", "src\Databases\Connectors\src\LangChain.Databases.Connectors.csproj", "{A7CF79E2-4DA4-417D-88AD-669B6FA717DB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LangChain.Databases.SemanticKernel", "src\Databases\SemanticKernel\src\LangChain.Databases.SemanticKernel.csproj", "{D1D3D997-AD91-45E6-9610-8C5080750CB0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -737,10 +737,10 @@ Global
 		{259C2E5D-6AA9-4384-8F55-24C4D4F1CF18}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{259C2E5D-6AA9-4384-8F55-24C4D4F1CF18}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{259C2E5D-6AA9-4384-8F55-24C4D4F1CF18}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A7CF79E2-4DA4-417D-88AD-669B6FA717DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A7CF79E2-4DA4-417D-88AD-669B6FA717DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A7CF79E2-4DA4-417D-88AD-669B6FA717DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A7CF79E2-4DA4-417D-88AD-669B6FA717DB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D1D3D997-AD91-45E6-9610-8C5080750CB0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D1D3D997-AD91-45E6-9610-8C5080750CB0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D1D3D997-AD91-45E6-9610-8C5080750CB0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D1D3D997-AD91-45E6-9610-8C5080750CB0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -887,8 +887,8 @@ Global
 		{6C053245-EB8B-4C98-85AC-B00BA222ABBB} = {3FCA798D-2CBE-41E7-B5F7-65B4AF6EBC8A}
 		{3CB20A37-5E76-40D1-91F6-47C96FA06AE5} = {A098FF69-D8B5-4B2B-83D5-F777D3817F15}
 		{259C2E5D-6AA9-4384-8F55-24C4D4F1CF18} = {3CB20A37-5E76-40D1-91F6-47C96FA06AE5}
-		{FB278634-1B3C-4AF1-BFC1-BCC7E07D1388} = {A098FF69-D8B5-4B2B-83D5-F777D3817F15}
-		{A7CF79E2-4DA4-417D-88AD-669B6FA717DB} = {FB278634-1B3C-4AF1-BFC1-BCC7E07D1388}
+		{3F426832-2CF3-4921-819F-13342F69A6E6} = {A098FF69-D8B5-4B2B-83D5-F777D3817F15}
+		{D1D3D997-AD91-45E6-9610-8C5080750CB0} = {3F426832-2CF3-4921-819F-13342F69A6E6}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {5C00D0F1-6138-4ED9-846B-97E43D6DFF1C}
