--- src/Databases/Connectors/src/LangChain.Databases.Connectors.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-	<NoWarn>$(NoWarn);SKEXP0020;SKEXP0001;CS3001</NoWarn>
-  </PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Microsoft.SemanticKernel.Core" />
-		<PackageReference Include="System.Linq.Async" />
-		<PackageReference Include="System.Linq.Async" />
-		<ProjectReference Include="..\..\..\Utilities\Pollyfils\src\LangChain.Polyfills.csproj" />
-		<ProjectReference Include="..\..\Abstractions\src\LangChain.Databases.Abstractions.csproj" />
-	</ItemGroup>
-
-</Project>
