+++ src/Databases/Milvus/src/LangChain.Databases.Milvus.csproj
@@ -7,7 +7,7 @@
 
     <ItemGroup>
 		<PackageReference Include="Microsoft.SemanticKernel.Connectors.Milvus" />
-        <ProjectReference Include="..\..\Connectors\src\LangChain.Databases.Connectors.csproj" />
+        <ProjectReference Include="..\..\SemanticKernel\src\LangChain.Databases.SemanticKernel.csproj" />
     </ItemGroup>
 
 </Project>
