+++ src/Databases/Qdrant/src/QdrantVectoDatabase.cs
@@ -1,4 +1,4 @@
-﻿using LangChain.Databases.Connectors;
+﻿using LangChain.Databases.SemanticKernel;
 using Microsoft.SemanticKernel.Connectors.Qdrant;
 
 namespace LangChain.Databases.Qdrant
