+++ src/Databases/Redis/src/RedisVectorCollection.cs
@@ -1,4 +1,4 @@
-﻿using LangChain.Databases.Connectors;
+﻿using LangChain.Databases.SemanticKernel;
 using Microsoft.SemanticKernel.Connectors.Redis;
 
 namespace LangChain.Databases.Redis
