+++ src/Databases/Weaviate/src/WeaviateVectorDatabase.cs
@@ -1,4 +1,4 @@
-﻿using LangChain.Databases.Connectors;
+﻿using LangChain.Databases.SemanticKernel;
 using Microsoft.SemanticKernel.Connectors.Weaviate;
 
 namespace LangChain.Databases.Weaviate
